requirements.txt
numpy
pandas
pybind11
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
requests
wikipedia
langchain-text-splitters
langchain-openai
chromadb
langchain-ollama
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain_classic
streamlit
langchain-groq
fastApi
uvicorn
langserve
sse_starlette
duckduckgo-search
ddgs
SQLAlchemy
mysql-connector-python
mysql
validators == 0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
cassio 
datasets
tiktoken
PyPDF2
gradio